[FIX] entrypoint: use /8 as range for private ip addresses
The subnet 10.89.0.0/16 can be extended to 10.0.0.0/8,
as the whole subnet is defined as private addressess by RFC 1918

podman can use 10.0.2.2 to resolve host.containers.internal,
the ip from the container to the host,
and hence it must be allowed to be able
to access mailhog from the odoo container to the mailhog
port 1025 on the host.

Private IP Address Ranges (RFC 1918)
10.0.0.0 to 10.255.255.255 (10.0.0.0/8)
172.16.0.0 to 172.31.255.255 (172.16.0.0/12)
192.168.0.0 to 192.168.255.255 (192.168.0.0/16)

Extend the same strategy for 127.0.0.0,
to be uniform.

The loopback range is defined by RFC 1122 and includes all addresses from 127.0.0.0 to 127.255.255.255
So 127.0.0.0/8